pyproject.toml
[project]
name = "carv-os-scoring"
version = "0.1.0"
description = "Candidate scoring service using Google Gemini"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-pubsub>=2.27.0",
    "google-genai>=1.0.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-gcp-trace>=1.8.0",
    "opentelemetry-exporter-gcp-monitoring>=1.8.0a0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scoring"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
